% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{generate_flat_code}
\alias{generate_flat_code}
\alias{generate_flat_code_server}
\alias{generate_flat_code_ui}
\title{Code generation plugin module}
\usage{
generate_flat_code(
  server = generate_flat_code_server,
  ui = generate_flat_code_ui
)

generate_flat_code_server(id, board, ...)

generate_flat_code_ui(id, board)
}
\arguments{
\item{server, ui}{Server/UI for the plugin module}

\item{id}{Namespace ID}

\item{board}{Reactive values object}

\item{...}{Extra arguments passed from parent scope}
}
\value{
A plugin container inheriting from \code{generate_code} is returned by
\code{generate_code()}, while the UI component (e.g. \code{generate_code_ui()}) is
expected to return shiny UI (i.e. \code{\link[shiny:reexports]{shiny::tagList()}}) and the server
component (i.e. \code{generate_code_server()}) is expected to return \code{NULL}.
}
\description{
All code necessary for reproducing a data analysis as set up in blockr can
be made available to the user. While the \code{generate_code} plugin bundled with
blockr.core can provide code to accurately reproduce all results (final and
intermediate), this plugin is aimed at making this exported code as idiomatic
as possible.
}
